# laravel-openapi-demo/public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Laravel OpenAPI Demo API",
        "version": "1.0.0",
        "description": null,
        "contact": {
            "name": null,
            "email": null,
            "url": null
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Direction to sort (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "default": "asc"
                        }
                    }
                ],
                "requestBody": null,
                "responses": {
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "A paginated list of resources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "format": "uri"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "format": "uri"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "nullable": true
                                                }
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "format": "uri"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "tasks.index",
                "summary": "List all tasks",
                "description": "Returns a list of task resources.",
                "tags": [
                    "Tasks"
                ]
            },
            "post": {
                "parameters": [],
                "requestBody": {
                    "description": "Create a new task",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "Task name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "Task description"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "email": [
                                                    "The email must be a valid email address."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "tasks.store",
                "summary": "Create a new task",
                "description": "Creates a new task resource.",
                "tags": [
                    "Tasks"
                ]
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": null,
                "responses": {
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "The requested resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "tasks.show",
                "summary": "Get a specific task",
                "description": "Returns a specific task resource.",
                "tags": [
                    "Tasks"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update an existing task",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "Task name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "Task description"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "tasks.update",
                "summary": "Update a task",
                "description": "Updates an existing task resource.",
                "tags": [
                    "Tasks"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": null,
                "responses": {
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Resource deleted successfully"
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "tasks.destroy",
                "summary": "Delete a task",
                "description": "Deletes a task resource.",
                "tags": [
                    "Tasks"
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Task": {
                "type": "object",
                "title": "Task",
                "description": "Represents a task item in the to-do list.",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Title"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description"
                    },
                    "status": {
                        "type": "string",
                        "description": "Status"
                    },
                    "due_date": {
                        "type": "string",
                        "description": "Due date",
                        "format": "date"
                    },
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "description": "Created at",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "description": "Updated at",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "status"
                ]
            }
        },
        "securitySchemes": [],
...
    },
    "security": [],
...
}